Added new function (Doughnut), added opacity function in TraitGraph
I added a new graph for relative score comparisons. It is likely most used for voting where we want it to be a percentage rather than a raw vote count. This still needs work to manage exceptions cases and light colors, but works.

Added in the ability to change the alpha levels of the bars in the TraitGraph if we want them to look more solid.